pwm: Convert period and duty to unsigned int
We don't need a 64 bits value to store nanoseconds

Discused with:	ian, jhibbits


git-svn-id: svn+ssh://svn.freebsd.org/base/head@342087 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/dev/pwm/pwm_if.m
@@ -46,8 +46,8 @@ METHOD device_t get_bus {
 METHOD int channel_config {
 	device_t dev;
 	int channel;
-	uint64_t period;
-	uint64_t duty;
+	unsigned int period;
+	unsigned duty;
 };
 
 #
@@ -57,8 +57,8 @@ METHOD int channel_config {
 METHOD int channel_get_config {
 	device_t dev;
 	int channel;
-	uint64_t *period;
-	uint64_t *duty;
+	unsigned int *period;
+	unsigned int *duty;
 };
 
 #
